mobisec/mobisec_project/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
certifi==2023.5.7
charset-normalizer==3.1.0
constantly==15.1.0
Django==4.2.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==10.0.0
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==2.0.4
zope.interface==6.0